Design Gallery Layout
The user will pick one of three places to visit at each stage.  If no
image is present, the current fallback is the application icon.  The
name of the place is to be displayed below the image.  If the user has
their own plans (use case example: have brought packed lunch with them,
so don't need location for lunch), they have the option to ignore that
session.