Very Basic FullInfo implementation
Added a very basic version of FullInfo with dummy data and buttons that
don't do anything.  Good enough for a demonstration.  The next stage is
a big Internet one, although I can add support for actually getting data
to display based on intents before the first commit to master for that.
The layout of the news feed in particular will change when formatting
for specific sources as well as potential links are introduced.