Route and FullInfo finalised functionality
Primary change as far as others care is IRoute and Route, which
represent the current route.  There's a static object in
MainScreenActivity for this.
Other changes are to FullInfoActivity (and related):  The news feed now
gets the place's specific handle, the image is being obtained and
displayed and it now works on slightly lower resolutions (yaaaaay)
Also buttons.  The buttons are done.  Only need now to improve news feed
look and add reference to Foursquare page